add a ranking constant that further penalizes stories with little data
this is because although the estimated quality for each story might be accurate, it is not perfectly precise for
stories where we don't have a lot of data. So we will over-estimate the quality of some stories, and if there is a large number of average-quality stories, some portion of them will be over-estimated and end up on the front page just by chance.

Adding a penalty that further penalizes low-quality stories. This is just a wild guess: I need to do further estimate to find
the optimal penalty.