shutdown http server on http handler panic
Although the http server recovers from panics, I believe it is now considered best practice to terminate
(and restart) the process if there is a panic, because as a general rule a panic means program integrity is no longer certain
and continued execution can be unsafe.

A clean shutdown should allow current HTTP requests to terminate.

Ideally, it should also cleanly shut down the crawler. I think this means
that if the crawler has already finished fetching data from HN but has not yet finished inserting into the DB, it should finish before exiting. But if it has not yet
started inserting, it should stop the crawl and exit immediately. This is a future todo.