Estimate submission time of resubmitted stories, estimate and apply penalties. Refactor logic for updating qnRank field.
We now use SQL statements under the sql/ dir to update 3 columns: 1) submissionTime (for resubmitted stories) 2) penalty, and 3) qnRank.

Updating qnRank as a SQL statement is cleaner since we no longer do this as a side-effect of generating the home page. These columns must
be computed in order.

Fixes #39, #47 and #56