make template names match template file name including extension
in the previous commit to allow one template per file, the extension was stripped from the template name. But
this means there could be two templates with the same name (e.g. foo.css.tmpl and foo.html.tmpl). Using the full
file name as the template name eliminates this possibility.