Fix data quality issues due to gaps in crawl data
Various data quality issues are being caused by situations where there are gaps in the
dataset for a story (two datapoints for a story are not two contiguous minutes). This can happen
if the story's rank temporarily dips below 90, or if the crawler stops running for a while.

This was causing various data quality issues: the submissionTime and penalty calculations
were "forgetting" the previously calculated penalty or submissionTime (since they were only looking at a 1-hour window),
the cumulativeUpvotes calculation was often two high (since it accumulated all upvotes that occurred during a gap in
the one minute after the gap).