Refactor plots to include plot data JSON in initial request
The immediate reason for this was to make all plots have the same x-axis range, so that a datapoint in one plot
aligned with datapoints for the same timestamp in other plots. This required getting the x-axis range from the plot
data for one plot before plotting the others. The async code for fetching plot data complicated this, and the code is
now simpler and faster.