Ranking formula uses age^upvoteRate in denominator
Here's my comment from Whatsapp justifying this:

I also think we can improve our ranking formula in a way that will improve overall upvote rate. Now we substitute  for  in the ranking formula, with the idea that
a story's score should grow roughly with age, but the score growing faster for story's with a high upvote rate. But I think that this relationship should be super-linear: stories with high upvote
rate should both get more attention, and then get more upvotes per unit of attention. So maybe something like . We might be able to figure out what this curve should look like with some analysis in R.