Implement Thomson Sampling!
We are Bayesians who believe the upvoteRate is a poisson distribution, and our beliefs about the parameter of that distribution is a gamma distribution.
That distribution has alpha=cumulativeUpvotes + overallPriorWeight, and beta approximately equal to cumulativeExpectedUpvotes + overallPriorWeight, but using the fatigue formula.

Intead of just taking the mean of this distribution, we sample from the distribution, and ues this as our upvoteRate in the scoring formula. In this way, our scoring formula ranks
stories highly if they have a high upvote rate, or if there is a lot of uncertainty about the upvote rate (low sample size), in I believe a theoretically optimal way.