Refactor archiving and purging
Problem was some stories had so many records in the dataset that generating the archive file alone was taking more than a minute.
Deleting was also taking more than a minute. We were trying to fit all archiving/purging operations in the time between crawls.

So now we do the archiving/purging in a separate process. Generating and uploading archive files can happen concurrently with the crawl.
But the purging, which involves deletes, has to happen between crawls. Use a channel to signal the idle period during which purging can happen.